+++ src/main/java/subway/dao/LineDao.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
-import subway.domain.Line;
+import subway.domain.subway.Line;
 
 @Repository
 public class LineDao {
