+++ src/main/java/subway/domain/subway/Sections.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.subway;
 
 import java.util.ArrayList;
 import java.util.HashMap;
